_OLD_BACKUP/python_core/sync_and_go.py
import logging
import os
import psycopg2
import requests
from dotenv import load_dotenv

# Load Environment Variables from Root
load_dotenv(os.path.join(os.path.dirname(os.path.dirname(__file__)), '.env'))

# Configuración de Logs Sovereign
logging.basicConfig(
    level=logging.INFO,
    format='%(asctime)s - [SYNC & GO] - %(message)s'
)
logger = logging.getLogger(__name__)

class SyncMaster:
    """
    🔗 SYNC & GO MASTER SCRIPT
    Finalizes production readiness: Backend Data -> Frontend Availability.
    """
    
    def __init__(self):
        self.db_url = os.getenv("DATABASE_URL")
        self.frontend_url = "http://localhost:3000" # Or production URL

    def check_database_readiness(self):
        """Checks if KnowledgeKernel has enough data for launch."""
        try:
            conn = psycopg2.connect(self.db_url)
            cur = conn.cursor()
            cur.execute("SELECT COUNT(*) FROM knowledge_kernel WHERE is_premium = TRUE")
            count = cur.fetchone()[0]
            conn.close()
            
            logger.info(f"📊 DB Check: {count} Premium Assets ready.")
            return count > 0
        except Exception as e:
            logger.error(f"❌ DB Check Failed: {e}")
            return False

    def check_frontend_connectivity(self):
        """Simulates a health check to the Frontend."""
        # For local dev, we might verify if the server is up, or just assume success if code is deployed.
        # This is a placeholder for a real health check endpoint.
        logger.info("🌐 Frontend Connectivity: ASSUMED ACTIVE (Deployment pending).")
        return True

    def execute_sync(self):
        logger.info("🚀 EXECUTING SYNC & GO PROTOCOL...")
        
        db_ready = self.check_database_readiness()
        fe_ready = self.check_frontend_connectivity()
        
        if db_ready and fe_ready:
            logger.info("✅ SYNC COMPLETE. SYSTEM IS GREEN.")
            print("\n" + "="*50)
            print("   MISPLANESCR SYSTEM STATUS: OPERATIONAL")
            print("   MODE: KAIZEN ∞ (Sovereign)")
            print("   MONETIZATION: ACTIVE (SINPE 60906359)")
            print("="*50 + "\n")
        else:
            logger.error("❌ SYNC FAILED. Data or Frontend not ready.")

if __name__ == "__main__":
    master = SyncMaster()
    master.execute_sync()
