requirements.txt
langchain
langchain-community
langchain-chroma
langchain-openai
langchain-google-genai
chromadb
pypdf
python-dotenv
openai
tiktoken
psycopg2-binary
langchain-postgres
pgvector
